greencloud/requirement.txt
<<<<<<< HEAD
mysqlclient
requests
django
django-rest-framework
django-cors-headers
djangorestframework-simplejwt
bs4
=======
mysqlclient
requests
django
django-rest-framework
django-cors-headers
djangorestframework-simplejwt
>>>>>>> origin/main
